// src/app/modules/core/dmo/model/dmo.ts
export class CreditDebitTable {
  transactionsName: string;
  credit: string;
  debit: string;
}

export class TbInvestRediscountTable {
  memoNo: string;
  adviceNo: string;
  adviceDate: string;
  adviceBy: string;
  transaction: string;
  description: string;
  credit: string;
  debit: string;
}

export class LoanReceivedPaidDataSourceTable {
  memoNo: string;
  adviceNo: string;
  adviceDate: string;
  adviceBy: string;
  transaction: string;
  description: string;
  credit: string;
  debit: string;
}

export class WaysAndMeansAdvancesDataSourceTable {
  memoNo: string;
  adviceNo: string;
  adviceDate: string;
  adviceBy: string;
  transaction: string;
  description: string;
  credit: string;
  debit: string;
}

export class OtherIgaTransactionsDataSourceTable {
  memoNo: string;
  adviceNo: string;
  adviceDate: string;
  adviceBy: string;
  transaction: string;
  description: string;
  credit: string;
  debit: string;
}